Add analysis scripts and clean up experiments
- Add 5 new analysis scripts for trait vectors
- Delete obsolete markdown planning files
- Update documentation
- Clean up cognitive_load experiment (removed)
- Fix layer 26 references (Gemma 2B has 26 layers, 0-25)

+++ traitlens/methods.py
@@ -204,8 +204,8 @@ class ProbeMethod(ExtractionMethod):
                 "Install with: pip install scikit-learn"
             )
 
-        # Prepare data
-        X = torch.cat([pos_acts, neg_acts], dim=0).cpu().numpy()
+        # Prepare data (convert bfloat16 to float32 for numpy compatibility)
+        X = torch.cat([pos_acts, neg_acts], dim=0).to(torch.float32).cpu().numpy()
         y = np.concatenate([
             np.ones(pos_acts.shape[0]),
             np.zeros(neg_acts.shape[0])
